+++ matchms/filtering/filter_utils/derive_precursor_mz_and_parent_mass.py
@@ -2,7 +2,7 @@ import logging
 from matchms.constants import PROTON_MASS
 from matchms.filtering.repair_adduct.clean_adduct import (_clean_adduct,
                                                           load_adducts_dict)
-from matchms.filtering.repair_adduct.interpret_unknown_adduct import \
+from matchms.filtering.filter_utils.interpret_unknown_adduct import \
     get_multiplier_and_mass_from_adduct
